+++ effectful-plugin/CHANGELOG.md
@@ -1,5 +1,6 @@
-# effectful-plugin-2.0.0.2 (????-??-??)
+# effectful-plugin-2.1.0.0 (????-??-??)
 * Drop support for GHC < 9.6.
+* Consider built-in instances when filtering candidates.
 
 # effectful-plugin-2.0.0.1 (2025-08-30)
 * Small optimization of checking suitable effects for unification.